librerias y animaciones final

+++ scripts/taks.js
@@ -4,9 +4,10 @@ if (!localStorage.jwt) {
   location.replace('./index.html');
 }
 
-
 /* ------ comienzan las funcionalidades una vez que carga el documento ------ */
 window.addEventListener('load', function () {
+  /* ------------------------- iniciamos libreria AOS ------------------------- */
+  AOS.init();
 
   const urlTareas = 'https://ctd-fe2-todo.herokuapp.com/v1/tasks';
   const urlUsuario = 'https://ctd-fe2-todo.herokuapp.com/v1/users/getMe';
@@ -25,12 +26,32 @@ window.addEventListener('load', function () {
   /* -------------------------------------------------------------------------- */
 
   btnCerrarSesion.addEventListener('click', function () {
-    const cerrarSesion = confirm("¿Desea cerrar sesión?");
-    if (cerrarSesion) {
-      //limpiamos el localstorage y redireccioamos a login
-      localStorage.clear();
-      location.replace('./index.html');
-    }
+    // const cerrarSesion = confirm("¿Desea cerrar sesión?");
+    // if (cerrarSesion) {
+    //   //limpiamos el localstorage y redireccioamos a login
+    //   localStorage.clear();
+    //   location.replace('./index.html');
+    // }
+    Swal.fire({
+      title: '¿Desea cerrar sesión?',
+      icon: 'info',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Confirmar',
+      cancelButtonText: 'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          '¡Hasta luego!',
+          'Te esperamos pronto.',
+          'success'
+        );
+        localStorage.clear();
+        location.replace('./index.html');
+      }
+    });
+
   });
 
   /* -------------------------------------------------------------------------- */
@@ -142,7 +163,7 @@ window.addEventListener('load', function () {
         contador++;
         //lo mandamos al listado de tareas completas
         tareasTerminadas.innerHTML += `
-          <li class="tarea">
+          <li class="tarea" data-aos="fade-up">
             <div class="hecha">
               <i class="fa-regular fa-circle-check"></i>
             </div>
@@ -158,7 +179,7 @@ window.addEventListener('load', function () {
       } else {
         //lo mandamos al listado de tareas sin terminar
         tareasPendientes.innerHTML += `
-          <li class="tarea">
+          <li class="tarea" data-aos="fade-down">
             <button class="change" id="${tarea.id}"><i class="fa-regular fa-circle"></i></button>
             <div class="descripcion">
               <p class="nombre">${tarea.description}</p>
@@ -227,22 +248,41 @@ window.addEventListener('load', function () {
     btnBorrarTarea.forEach(boton => {
       //a cada boton de borrado le asignamos la funcionalidad
       boton.addEventListener('click', function (event) {
-        const id = event.target.id;
-        const url = `${urlTareas}/${id}`
+        Swal.fire({
+          title: '¿Confirma eliminar la tarea?',
+          icon: 'question',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Confirmar',
+          cancelButtonText: 'Cancelar'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            /* -------------------- disparamos el fetch para eliminar ------------------- */
+            const id = event.target.id;
+            const url = `${urlTareas}/${id}`
 
-        const settingsCambio = {
-          method: 'DELETE',
-          headers: {
-            "Authorization": token,
+            const settingsCambio = {
+              method: 'DELETE',
+              headers: {
+                "Authorization": token,
+              }
+            }
+            fetch(url, settingsCambio)
+              .then(response => {
+                console.log("Borrando tarea...");
+                console.log(response.status);
+                //vuelvo a consultar las tareas actualizadas y pintarlas nuevamente en pantalla
+                consultarTareas();
+              })
+
+            Swal.fire(
+              'Tarea eliminada.',
+            );
+
           }
-        }
-        fetch(url, settingsCambio)
-          .then(response => {
-            console.log("Borrando tarea...");
-            console.log(response.status);
-            //vuelvo a consultar las tareas actualizadas y pintarlas nuevamente en pantalla
-            consultarTareas();
-          })
+        });
+
       })
     });
   }
